Fixed: Drag and drop sometimes moves tab to bottom
https://github.com/mbnuqw/sidebery/issues/528

Not everyone can reproduce, but it happens at random ~30% of the time
for me while reordering tabs.

Problem Analysis: (to explain why the fix helps)

When dragging crosses the border between two tabs, we shortly
enter/leave a tiny "Nowhere" zone. (Events OnDragEnter and OnDragMove
targeting the outer <div> of the tab, which does not have a
data-dnd-type attribute, followed by the same events targeting the
correct inner element again.)

Before the bugfix, we updated the position during DragMove event while
in that zone. There is a check (line 693) that sets the index to -1 when
we are not in a "Tab" zone. Because the global position was already
updated, the later events are not recognized as a position change, so we
do not reach the same line of code again when back in the "Tab" zone.

With the index -1 the tab will be dropped at the end, instead of the
cursor position.